+++ tests/pytorch/test_fusible_ops.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from collections.abc import Iterable, Sequence
 import functools
 import io
+import os
 import math
 import random
 from typing import Optional
@@ -42,7 +43,6 @@ from transformer_engine.pytorch import (
 )
 from transformer_engine.pytorch.tensor.grouped_tensor import GroupedTensor
 from transformer_engine.pytorch.cpp_extensions.gemm import general_grouped_gemm_for_grouped_tensor
-from transformer_engine.pytorch.module.base import get_dummy_wgrad
 import transformer_engine_torch as tex
 
 # Import utility functions
@@ -51,6 +51,7 @@ from utils import (
     assert_close_grads,
     dtype_tols,
     make_recipe,
+    MegatronTrainingHelper,
     quantization_tols,
     reset_rng_states,
 )
@@ -210,76 +211,6 @@ def to_cpu(tensor: Optional[torch.Tensor]) -> Optional[torch.Tensor]:
     out = out.to(dtype=torch.float64, device="cpu")
     out = out.requires_grad_(requires_grad=tensor.requires_grad)
     return out
-
-
-class MegatronTrainingHelper:
-    """Test-side stand-in for the Megatron-Core DDP / MegatronFSDP wrapper.
-    Megatron's DDP wrapper (and MegatronFSDP) owns the per-parameter
-    ``main_grad`` buffer and the ``overwrite_main_grad`` /
-    ``grad_added_to_main_grad`` attributes that coordinate
-    ``fuse_wgrad_accumulation`` with TE modules. These helpers reproduce the
-    relevant slice of that protocol so TE tests can exercise the
-    accumulate-into-``main_grad`` code path without pulling in the full
-    Megatron-Core dependency.
-    """
-
-    @staticmethod
-    def init_main_grad_buffers(
-        weight_params: Iterable[torch.nn.Parameter],
-        *,
-        fill_value: float,
-        overwrite_main_grad: bool,
-        zero_out_wgrad: bool = False,
-        dtype: torch.dtype = torch.float32,
-    ) -> None:
-        """Allocate ``main_grad`` and stamp the wrapper attributes on each
-        param, mirroring what the Megatron DDP/FSDP wrapper does before
-        backward."""
-        for wp in weight_params:
-            wp.main_grad = torch.full(wp.size(), fill_value, device=wp.device, dtype=dtype)
-            wp.overwrite_main_grad = overwrite_main_grad
-            wp.zero_out_wgrad = zero_out_wgrad
-            wp.grad_added_to_main_grad = False
-
-    @staticmethod
-    def verify_main_grad_accumulation(
-        weight_params: Iterable[torch.nn.Parameter],
-        *,
-        expected_main_grads: Iterable[torch.Tensor],
-        rtol: float = 0.0,
-        atol: float = 0.0,
-    ) -> None:
-        """Check that backward produced what the Megatron wrapper expects:
-        each ``main_grad`` matches ``expected_main_grads``,
-        ``grad_added_to_main_grad`` was flipped to ``True`` so the wrapper's
-        post-backward hooks won't double-accumulate, and ``param.grad`` was
-        replaced by the cached dummy tensor (so a wrapper hook that did
-        ``main_grad += grad`` would be a no-op rather than double-counting).
-        """
-        for wp, expected in zip(weight_params, expected_main_grads):
-            torch.testing.assert_close(wp.main_grad.to(expected), expected, rtol=rtol, atol=atol)
-
-            assert wp.grad_added_to_main_grad is True, (
-                "weight.grad_added_to_main_grad was not flipped to True; "
-                "the Megatron DDP/FSDP wrapper hook will double-accumulate."
-            )
-
-            # ``.grad`` should be the cached dummy tensor returned by
-            # ``get_dummy_wgrad`` -- shared storage, not the real wgrad.
-            expected_dummy = get_dummy_wgrad(list(wp.size()), wp.dtype)
-            assert (
-                wp.grad is not None
-            ), "weight.grad is None; the Megatron protocol expects a dummy tensor stand-in here."
-            assert wp.grad.data_ptr() == expected_dummy.data_ptr(), (
-                "weight.grad does not share storage with the cached dummy "
-                "wgrad; downstream wrapper hooks risk double-accumulating."
-            )
-            if getattr(wp, "zero_out_wgrad", False):
-                assert torch.all(wp.grad == 0), (
-                    "weight.zero_out_wgrad=True but the dummy weight.grad "
-                    "was not zeroed; downstream hooks reading .grad would "
-                    "see stale bytes from the previous step."
-                )
 
 
 class TestSequentialContainer:
@@ -2119,7 +2050,8 @@ class TestBasicOps:
         single_grouped_bias: bool,
     ) -> None:
         """Grouped GEMM"""
-
+        if os.environ.get("NVTE_GROUPED_LINEAR_SINGLE_PARAM", "0") == "0" and (single_grouped_weight or single_grouped_bias):
+            pytest.skip("single_grouped_weight/single_grouped_bias requires NVTE_GROUPED_LINEAR_SINGLE_PARAM=1")
         # Split sizes
         split_sizes = [split_alignment * i for i in range(group_size)]
         random.shuffle(split_sizes)
@@ -2321,7 +2253,8 @@ class TestBasicOps:
         Exercises the grouped-tensor / cublas-grouped-gemm path which uses
         GPU-resident split offsets and is the only flow safe to capture.
         """
-
+        if os.environ.get("NVTE_GROUPED_LINEAR_SINGLE_PARAM", "0") == "0" and (single_grouped_weight or single_grouped_bias):
+            pytest.skip("single_grouped_weight/single_grouped_bias requires NVTE_GROUPED_LINEAR_SINGLE_PARAM=1")
         if torch.cuda.get_device_capability() < (10, 0):
             pytest.skip("Grouped GEMM CUDA-graph-safe path requires SM100+ (Blackwell)")
         # Skip invalid configurations
